Fix output of sample data
